// taskflow-graphql/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
  [P in K]-?: NonNullable<T[P]>;
};
// ...
export type Board = {
  __typename?: 'Board';
  columns: Array<BoardColumn>;
  completedTaskCount: Scalars['Int']['output'];
  createdAt: Scalars['DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  isShared: Scalars['Boolean']['output'];
  name: Scalars['String']['output'];
  settings: BoardSettings;
  taskCount: Scalars['Int']['output'];
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export enum MarkdownFormat {
  GithubFlavored = 'GITHUB_FLAVORED',
  Obsidian = 'OBSIDIAN',
  Standard = 'STANDARD',
}
// ...
export type Query = {
  __typename?: 'Query';
  aiSuggestedTasks: Array<Task>;
  board?: Maybe<Board>;
  boards: Array<Board>;
  currentBoard?: Maybe<Board>;
  exportBoardAsMarkdown: Scalars['String']['output'];
  exportTaskAsMarkdown: Scalars['String']['output'];
  exportTasksAsMarkdown: Scalars['String']['output'];
  label?: Maybe<Label>;
  labels: Array<Label>;
  nextRecommendedTask?: Maybe<Task>;
  searchTasksByNaturalLanguage: Array<Task>;
  task?: Maybe<Task>;
  taskStatistics: TaskStatistics;
  tasks: Array<Task>;
  template?: Maybe<Template>;
  templates: Array<Template>;
  webhook?: Maybe<Webhook>;
  webhookDeliveries: Array<WebhookDelivery>;
  webhookDelivery?: Maybe<WebhookDelivery>;
  webhookStats: WebhookStats;
  webhooks: Array<Webhook>;
};

export type QueryAiSuggestedTasksArgs = {
  context: AiContextInput;
};

export type QueryBoardArgs = {
  id: Scalars['ID']['input'];
};

export type QueryExportBoardAsMarkdownArgs = {
  boardId: Scalars['ID']['input'];
  filters?: InputMaybe<TaskFilters>;
};
// ...
export type TaskFilters = {
  dueAfter?: InputMaybe<Scalars['DateTime']['input']>;
  dueBefore?: InputMaybe<Scalars['DateTime']['input']>;
  labels?: InputMaybe<Array<Scalars['String']['input']>>;
  priority?: InputMaybe<Priority>;
  search?: InputMaybe<Scalars['String']['input']>;
  status?: InputMaybe<TaskStatus>;
};
// ...
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type ResolversParentTypes = {
  AIContextInput: Partial<AiContextInput>;
  AISuggestion: Partial<
    Omit<AiSuggestion, 'task'> & { task?: Maybe<ResolversParentTypes['Task']> }
  >;
  Attachment: AttachmentRecord;
  AttachmentInput: Partial<AttachmentInput>;
  Board: BoardRecord;
  BoardColumn: BoardColumnRecord;
  BoardColumnInput: Partial<BoardColumnInput>;
  BoardSettings: BoardSettingsRecord;
  BoardSettingsInput: Partial<BoardSettingsInput>;
  Boolean: Partial<Scalars['Boolean']['output']>;
  CreateBoardInput: Partial<CreateBoardInput>;
  CreateLabelInput: Partial<CreateLabelInput>;
  CreateSubTaskInput: Partial<CreateSubTaskInput>;
  CreateTaskInput: Partial<CreateTaskInput>;
  CreateTemplateInput: Partial<CreateTemplateInput>;
  CreateWebhookInput: Partial<CreateWebhookInput>;
  DateTime: Partial<Scalars['DateTime']['output']>;
  Float: Partial<Scalars['Float']['output']>;
  ID: Partial<Scalars['ID']['output']>;
  Int: Partial<Scalars['Int']['output']>;
  JSON: Partial<Scalars['JSON']['output']>;
  Label: LabelRecord;
  MarkdownMetadata: Partial<MarkdownMetadata>;
  MarkdownReport: Partial<MarkdownReport>;
  MarkdownReportInput: Partial<MarkdownReportInput>;
  Mutation: {};
  PriorityBreakdown: Partial<PriorityBreakdown>;
  Query: {};
  RecurrenceConfig: RecurrenceConfigRecord;
  RecurrenceConfigInput: Partial<RecurrenceConfigInput>;
  ScheduleConstraints: Partial<ScheduleConstraints>;
  ScheduleOptimization: Partial<
    Omit<ScheduleOptimization, 'optimizedTasks'> & {
      optimizedTasks: Array<ResolversParentTypes['Task']>;
    }
  >;
  StatusBreakdown: Partial<StatusBreakdown>;
  String: Partial<Scalars['String']['output']>;
  SubTask: SubTaskRecord;
  Subscription: {};
  SuggestedAction: Partial<SuggestedAction>;
  Task: TaskRecord;
  TaskFilters: Partial<TaskFilters>;
  TaskStatistics: Partial<TaskStatistics>;
  TaskTemplateData: TaskTemplateDataRecord;
  TaskTemplateDataInput: Partial<TaskTemplateDataInput>;
  Template: TemplateRecord;
  UpdateBoardInput: Partial<UpdateBoardInput>;
  UpdateLabelInput: Partial<UpdateLabelInput>;
  UpdateSubTaskInput: Partial<UpdateSubTaskInput>;
  UpdateTaskInput: Partial<UpdateTaskInput>;
  UpdateTemplateInput: Partial<UpdateTemplateInput>;
  UpdateWebhookInput: Partial<UpdateWebhookInput>;
  UserPreferencesInput: Partial<UserPreferencesInput>;
  Webhook: Partial<Webhook>;
  WebhookDelivery: Partial<WebhookDelivery>;
  WebhookStats: Partial<WebhookStats>;
  WorkingHoursInput: Partial<WorkingHoursInput>;
};
// ...
export type MutationResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation'],
> = {
  breakdownTask?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<MutationBreakdownTaskArgs, 'taskId'>
  >;
  createBoard?: Resolver<
    ResolversTypes['Board'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateBoardArgs, 'input'>
  >;
  createLabel?: Resolver<
    ResolversTypes['Label'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateLabelArgs, 'input'>
  >;
  createTask?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateTaskArgs, 'input'>
  >;
  createTaskFromNaturalLanguage?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateTaskFromNaturalLanguageArgs, 'query'>
  >;
  createTasks?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<MutationCreateTasksArgs, 'inputs'>
  >;
  createTemplate?: Resolver<
    ResolversTypes['Template'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateTemplateArgs, 'input'>
  >;
  createWebhook?: Resolver<
    ResolversTypes['Webhook'],
    ParentType,
    ContextType,
    RequireFields<MutationCreateWebhookArgs, 'input'>
  >;
  deleteBoard?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteBoardArgs, 'id'>
  >;
  deleteLabel?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteLabelArgs, 'id'>
  >;
  deleteTask?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteTaskArgs, 'id'>
  >;
  deleteTasks?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteTasksArgs, 'ids'>
  >;
  deleteTemplate?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteTemplateArgs, 'id'>
  >;
  deleteWebhook?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType,
    RequireFields<MutationDeleteWebhookArgs, 'id'>
  >;
  duplicateTask?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<MutationDuplicateTaskArgs, 'id'>
  >;
  generateMarkdownReport?: Resolver<
    ResolversTypes['MarkdownReport'],
    ParentType,
    ContextType,
    RequireFields<MutationGenerateMarkdownReportArgs, 'input'>
  >;
  moveTask?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<
      MutationMoveTaskArgs,
      'id' | 'targetColumnId' | 'targetPosition'
    >
  >;
  optimizeTaskSchedule?: Resolver<
    ResolversTypes['ScheduleOptimization'],
    ParentType,
    ContextType,
    RequireFields<MutationOptimizeTaskScheduleArgs, 'boardId'>
  >;
  restoreTask?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<MutationRestoreTaskArgs, 'id'>
  >;
  testWebhook?: Resolver<
    ResolversTypes['WebhookDelivery'],
    ParentType,
    ContextType,
    RequireFields<MutationTestWebhookArgs, 'id'>
  >;
  updateBoard?: Resolver<
    ResolversTypes['Board'],
    ParentType,
    ContextType,
    RequireFields<MutationUpdateBoardArgs, 'id' | 'input'>
  >;
  updateLabel?: Resolver<
    ResolversTypes['Label'],
    ParentType,
    ContextType,
    RequireFields<MutationUpdateLabelArgs, 'id' | 'input'>
  >;
  updateTask?: Resolver<
    ResolversTypes['Task'],
    ParentType,
    ContextType,
    RequireFields<MutationUpdateTaskArgs, 'id' | 'input'>
  >;
  updateTasks?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<MutationUpdateTasksArgs, 'ids' | 'input'>
  >;
  updateTemplate?: Resolver<
    ResolversTypes['Template'],
    ParentType,
    ContextType,
    RequireFields<MutationUpdateTemplateArgs, 'id' | 'input'>
  >;
  updateWebhook?: Resolver<
    ResolversTypes['Webhook'],
    ParentType,
    ContextType,
    RequireFields<MutationUpdateWebhookArgs, 'id' | 'input'>
  >;
};
// ...
export type QueryResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['Query'] = ResolversParentTypes['Query'],
> = {
  aiSuggestedTasks?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<QueryAiSuggestedTasksArgs, 'context'>
  >;
  board?: Resolver<
    Maybe<ResolversTypes['Board']>,
    ParentType,
    ContextType,
    RequireFields<QueryBoardArgs, 'id'>
  >;
  boards?: Resolver<Array<ResolversTypes['Board']>, ParentType, ContextType>;
  currentBoard?: Resolver<
    Maybe<ResolversTypes['Board']>,
    ParentType,
    ContextType
  >;
  exportBoardAsMarkdown?: Resolver<
    ResolversTypes['String'],
    ParentType,
    ContextType,
    RequireFields<QueryExportBoardAsMarkdownArgs, 'boardId'>
  >;
  exportTaskAsMarkdown?: Resolver<
    ResolversTypes['String'],
    ParentType,
    ContextType,
    RequireFields<QueryExportTaskAsMarkdownArgs, 'taskId'>
  >;
  exportTasksAsMarkdown?: Resolver<
    ResolversTypes['String'],
    ParentType,
    ContextType,
    RequireFields<QueryExportTasksAsMarkdownArgs, 'boardId'>
  >;
  label?: Resolver<
    Maybe<ResolversTypes['Label']>,
    ParentType,
    ContextType,
    RequireFields<QueryLabelArgs, 'id'>
  >;
  labels?: Resolver<
    Array<ResolversTypes['Label']>,
    ParentType,
    ContextType,
    Partial<QueryLabelsArgs>
  >;
  nextRecommendedTask?: Resolver<
    Maybe<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<QueryNextRecommendedTaskArgs, 'boardId'>
  >;
  searchTasksByNaturalLanguage?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<QuerySearchTasksByNaturalLanguageArgs, 'query'>
  >;
  task?: Resolver<
    Maybe<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    RequireFields<QueryTaskArgs, 'id'>
  >;
  taskStatistics?: Resolver<
    ResolversTypes['TaskStatistics'],
    ParentType,
    ContextType,
    Partial<QueryTaskStatisticsArgs>
  >;
  tasks?: Resolver<
    Array<ResolversTypes['Task']>,
    ParentType,
    ContextType,
    Partial<QueryTasksArgs>
  >;
  template?: Resolver<
    Maybe<ResolversTypes['Template']>,
    ParentType,
    ContextType,
    RequireFields<QueryTemplateArgs, 'id'>
  >;
  templates?: Resolver<
    Array<ResolversTypes['Template']>,
    ParentType,
    ContextType,
    Partial<QueryTemplatesArgs>
  >;
  webhook?: Resolver<
    Maybe<ResolversTypes['Webhook']>,
    ParentType,
    ContextType,
    RequireFields<QueryWebhookArgs, 'id'>
  >;
  webhookDeliveries?: Resolver<
    Array<ResolversTypes['WebhookDelivery']>,
    ParentType,
    ContextType,
    RequireFields<QueryWebhookDeliveriesArgs, 'webhookId'>
  >;
  webhookDelivery?: Resolver<
    Maybe<ResolversTypes['WebhookDelivery']>,
    ParentType,
    ContextType,
    RequireFields<QueryWebhookDeliveryArgs, 'id'>
  >;
  webhookStats?: Resolver<
    ResolversTypes['WebhookStats'],
    ParentType,
    ContextType
  >;
  webhooks?: Resolver<
    Array<ResolversTypes['Webhook']>,
    ParentType,
    ContextType
  >;
};